Normalized some of general controller errors to have stanardized error responses across multiple controllers

+++ app/controllers/stores_controller/StoreController.ts
@@ -1,14 +1,18 @@
 import { Request, Response } from "express";
 import { Types } from "mongoose";
-import Store, { IStore } from "../../models/Store";
-import StoreImage, { IStoreImage } from "../../models/StoreImage";
-import { IGenericController } from "../helpers/controllerInterfaces";
-import { RemoveResponse, resolveStoreItemImgDirectories, respondWithNotAllowedErr } from "../helpers/controllerHelpers"
-// helpers //
-import { respondWithDBError, respondWithInputError, resolveDirectoryOfImg, removeDirectoryWithFiles } from "../helpers/controllerHelpers";
-import StoreItem from "../../models/StoreItem";
-import StoreItemImage from "../../models/StoreItemImage";
+// models and model interfaces //
 import { IAdministrator } from "../../models/Administrator";
+import Store, { IStore } from "../../models/Store";
+import StoreItem from "../../models/StoreItem";
+import StoreImage, { IStoreImage } from "../../models/StoreImage";
+import StoreItemImage from "../../models/StoreItemImage";
+// additional generic interfaces //
+import { IGenericController } from "../helpers/controllerInterfaces";
+import { StoreData } from "./type_declarations/storesControllerTypes";
+// helpers //
+import { respondWithInputError, resolveDirectoryOfImg, removeDirectoryWithFiles } from "../helpers/controllerHelpers";
+import { RemoveResponse, resolveStoreItemImgDirectories, respondWithNotAllowedErr } from "../helpers/controllerHelpers"
+import { NotAllowedError, NotFoundError, processErrorResponse } from "../helpers/errorHandlers";
 
 interface IGenericStoreResponse {
   responseMsg: string;
@@ -18,11 +22,7 @@ interface IGenericStoreResponse {
   store?: IStore;
   stores?: IStore[];
 }
-export type StoreParams = {
-  title: string;
-  description: string;
-  images: IStoreImage[];
-}
+
 type StoreQueryPar = {
   title?: string;
   items?: string;
@@ -31,9 +31,6 @@ type StoreQueryPar = {
 }
 class StoresController implements IGenericController {
   getMany (req: Request, res: Response<IGenericStoreResponse>): Promise<Response> {
-    const { title, items, date, limit } : StoreQueryPar = req.query;
-    const queryLimit = limit ? parseInt(limit, 10) : 5;
-    
     const admin: IAdministrator = req.user as IAdministrator;
     // validate user and a business account set up //
     if (admin) {
@@ -43,6 +40,9 @@ class StoresController implements IGenericController {
     } else {
       return respondWithNotAllowedErr(res);
     }
+    // query option //
+    const { title, items, date, limit } : StoreQueryPar = req.query;
+    const queryLimit = limit ? parseInt(limit, 10) : 5;
     // custom queries //
     // sort by title alphabetically //
     if (title) {
@@ -59,7 +59,7 @@ class StoresController implements IGenericController {
         });
       })
       .catch((error) => {
-        return respondWithDBError(res, error);
+        return processErrorResponse(res, error);
       });
     }
     // sort by number of items //
@@ -77,7 +77,7 @@ class StoresController implements IGenericController {
         });
       })
       .catch((error) => {
-        return respondWithDBError(res, error);
+        return processErrorResponse(res, error);
       });
     }
     // sort by date created //
@@ -95,11 +95,11 @@ class StoresController implements IGenericController {
         });
       })
       .catch((error) => {
-        return respondWithDBError(res, error);
+        return processErrorResponse(res, error);
       });
     }
     // generic response //
-    return Store.find({})
+    return Store.find({ businessAccountId: admin.adminAccountId })
       .limit(queryLimit)
       .populate("images").exec()
       .then((stores) => {
@@ -109,12 +109,12 @@ class StoresController implements IGenericController {
         });
       })
       .catch((error) => {
-        return respondWithDBError(res, error);
+        return processErrorResponse(res, error);
       });
   }
   getOne (req: Request, res: Response<IGenericStoreResponse>): Promise<Response>  {
-    const _id: string = req.params._id;
     const admin: IAdministrator = req.user as IAdministrator;
+    const { _id : storeId } = req.params;
     // validate user and a business account set up //
     if (admin) {
       if (!admin.businessAccountId) {
@@ -123,47 +123,56 @@ class StoresController implements IGenericController {
     } else {
       return respondWithNotAllowedErr(res);
     }
+    // ensure a Store id was send //
+    if (storeId) return respondWithInputError(res, "Can't find store");
 
-    if (!_id) return respondWithInputError(res, "Can't find store");
-    return Store.findOne({ _id: _id })
-      .populate("images").exec()
+    return Store.findOne({ _id: storeId })
       .then((store) => {
         if (store) {
-          return res.status(200).json({
-            responseMsg: "Store found",
-            store: store
-          });
+          if ((store.businessAccountId as Types.ObjectId).equals(admin.businessAccountId!)) {
+            return store.populate("images").execPopulate();
+          } else {
+            throw new NotAllowedError({ messages: [ "Not allowed to view Stores belonging to another account" ] });
+          }
         } else {
-          return respondWithInputError(res, "Could not find store", 404);
+          throw new NotFoundError({ messages: [ "Requested Store was not found" ] });
         }
       })
+      .then((store) => {
+        return res.status(200).json({
+          responseMsg: "Store found",
+          store: store
+        });
+      })
       .catch((error) => {
-        return respondWithDBError(res, error);
+        return processErrorResponse(res, error);
       });
   }
   create (req: Request, res: Response<IGenericStoreResponse>): Promise<Response> {
-    const { title, description, images : storeImages }: StoreParams = req.body;
-    const imgIds: Types.ObjectId[] = [];
+    let businessAccountId: Types.ObjectId;
+    // first ensure admin and businessAccountId exist //
     const admin: IAdministrator = req.user as IAdministrator;
-    // validate user and a business account set up //
     if (admin) {
       if (!admin.businessAccountId) {
         return respondWithNotAllowedErr(res, "Action not allowed", 401, [ "Create a Business Acccount or be added to one to create Stores" ]);
+      } else {
+        businessAccountId = admin.businessAccountId;
       }
     } else {
       return respondWithNotAllowedErr(res);
     }
-    if (storeImages.length > 1) {
-      // let imgPromises: [Promise<Query<StoreImg>>];
-      for (const newImg of storeImages) {
-        imgIds.push(Types.ObjectId(newImg.url));
-      }
-    }
+    // store create arguments //
+    const { title, description }: StoreData = req.body;
+    
+
     const newStore = new Store({
+      businessAccountId: businessAccountId,
       title: title,
       description: description,
       numOfItems: 0,
-      images: [ ...imgIds ]
+      images: [],
+      createdAt: new Date(Date.now()),
+      editedAt: new Date(Date.now())
     });
 
     return newStore.save()
@@ -174,77 +183,112 @@ class StoresController implements IGenericController {
         });
       })
       .catch((err) => {
-        return respondWithDBError(res, err);
+        return processErrorResponse(res, err);
       });
   }
+
   edit (req: Request, res: Response<IGenericStoreResponse>): Promise<Response> {
-    const { _id } = req.params;
-    const { title, description, images : storeImages }: StoreParams = req.body;
+    const { _id : storeId } = req.params;
+    let businessAccountId: Types.ObjectId;
+    // first ensure admin and businessAccountId exist //
     const admin: IAdministrator = req.user as IAdministrator;
     // validate user and a business account set up //
     if (admin) {
       if (!admin.businessAccountId) {
         return respondWithNotAllowedErr(res, "Action not allowed", 401, [ "Create a Business Acccount or be added to one to edit Stores" ]);
+      } else {
+        businessAccountId = admin.businessAccountId;
       }
     } else {
       return respondWithNotAllowedErr(res);
     }
-    if (!_id) {
+    // ensure Store _id was sent //
+    if (!storeId) {
       return respondWithInputError(res, "Can't resolve store", 400);
     }
-
-    const updatesStoreImgs = storeImages.map((img) => Types.ObjectId(img._id));
-
-    return Store.findOneAndUpdate(
-      { _id: _id },
-      { 
-        $set: {
-          title: title,
-          description: description,
-          images: [ ...updatesStoreImgs ],
-          editedAt: new Date()
-        },
-      },
-      { new: true }
-     )
+    // Store update data //
+    const { title, description, images : storeImages = [] }: StoreData = req.body;
+    
+    return Store.findOne({ _id: storeId }).exec()
       .then((store) => {
-        return store?.populate("images")
+        if (!store) {
+          throw new NotFoundError({ messages: [ "Store to update not found" ] });
+        } else {
+          if ((store.businessAccountId as Types.ObjectId).equals(businessAccountId)) {
+            return Promise.resolve(store);
+          } else {
+            throw new NotAllowedError({ messages: [ "NOt allowed to edit a Store which does not belong to your account"] });
+          }
+        }
+      })
+      .then((store) => {
+        store.title = title;
+        store.description = description;
+        store.images = storeImages as Types.ObjectId[];
+        store.editedAt = new Date(Date.now());
+        return store.save();
+      })
+      .then((store) => {
+        return Store.populate(store, { path: "images" });
       })
       .then((store) => {
         return res.status(200).json({
           responseMsg: "Store Updates",
-          editedStore: store!
+          editedStore: store as IStore
         });
       })
       .catch((error) => {
         console.error(error);
-        return respondWithDBError(res, error);
+        return processErrorResponse(res, error);
       });
        
   }
 
-  delete (req: Request, res: Response<IGenericStoreResponse>): Promise<Response> {
-   const { _id : storeId } = req.params;
-   let numOfDeletedStoreImages: number; 
-   let numOfDeletedStoreItems: number;
-   let numOfDeletedStoreItemImages: number;
-   let storeItemIds: string[];
-
-   const admin: IAdministrator = req.user as IAdministrator;
+  delete (req: Request, res: Response<IGenericStoreResponse>) {
+    let numOfDeletedStoreImages: number; 
+    let numOfDeletedStoreItems: number;
+    let numOfDeletedStoreItemImages: number;
+    let storeItemIds: string[];
+    let businessAccountId: Types.ObjectId | string;
+    let storeToDelete: IStore;
+    // 
+    const { _id : storeId } = req.params;
+    const admin: IAdministrator = req.user as IAdministrator;
     // validate user and a business account set up //
     if (admin) {
       if (!admin.businessAccountId) {
         return respondWithNotAllowedErr(res, "Action not allowed", 401, [ "Create a Business Acccount or be added to one to delete and manage Stores" ]);
+      } else {
+        businessAccountId = admin.businessAccountId;
       }
     } else {
       return respondWithNotAllowedErr(res);
     }
+    // ensure Store model id was sent //
+    if (!storeId) {
+      return respondWithInputError(res, "Can't resolve Store");
+    }
 
-   if (!storeId) {
-     return respondWithInputError(res, "Can't find store");
-   }
-   else {
-    return StoreImage.find({ storeId: storeId })
+    return Store.findOne({ _id: storeId }).exec()
+      .then((store) => {
+        if (store) {
+          if ((store.businessAccountId as Types.ObjectId).equals(businessAccountId)) {
+            return store.populate("images").execPopulate();
+          } else {
+            throw new NotAllowedError({ messages: [ "Not allowed to delete Store which doesn't belong to your account" ] });
+          }
+        } else {
+          throw new NotFoundError({ messages: [ "Store to delete was not found" ] });
+        }
+      })
+      .then((store) => {
+        storeToDelete = store;
+        if (store.images.length > 0) {
+          return StoreImage.find({ storeId: storeId });
+        } else {
+          return Promise.resolve([]);
+        }
+      })
       .then((storeImages) => {
         if (storeImages.length > 0) {
           const imagesDirectory = resolveDirectoryOfImg(storeImages[0].absolutePath)
@@ -255,12 +299,20 @@ class StoresController implements IGenericController {
       }) 
       .then(({ success, numberRemoved, message }: RemoveResponse) => {
         // delete StoreImage models from DB //
-        return StoreImage.deleteMany({ storeId: storeId });
+        if (numberRemoved > 0) {
+          return StoreImage.deleteMany({ storeId: storeId });
+        } else {
+          return Promise.resolve({ n: 0 });
+        }
       })
       .then(({ n }) => {
         numOfDeletedStoreImages = n ? n : 0;
-        // find and delete StoreItems //
-        return StoreItem.find({ storeId: storeId });
+        // find and delete StoreItems if they are present //
+        if (storeToDelete.numOfItems > 0) {
+          return StoreItem.find({ storeId: storeId })
+        } else {
+          return Promise.resolve([]);
+        }
       })
       .then((storeItems) => {
         if (storeItems.length > 0) {
@@ -314,10 +366,11 @@ class StoresController implements IGenericController {
       })
       .catch((error) => {
         console.log(error)
-        return respondWithDBError(res, error);
+        return processErrorResponse(res, error);
       })
     }
-  }
+    
+  
 }
 
 export default StoresController;
